♻️ [REFACTOR] 기능 분리 및 MySQL 문으로 리팩토링
