+++ RoomCatcher_Data_AI/dataAnalyze/ver1/extract_tag_from_product.py
@@ -1,10 +1,10 @@
 # 매물의 이름과 설명, 기타 정보 등으로 키워드를 추출해내는 방법 
-
 import requests
-import sqlite3
+import pymysql  # MySQL 연결을 위해 pymysql 사용
 from django.core.exceptions import ImproperlyConfigured
 import os
 import json
+from dataAnalyze.common import connect_db  # 오타 수정
 
 def get_secret(setting, secrets):
     try:
@@ -12,7 +12,7 @@ def get_secret(setting, secrets):
     except KeyError:
         error_msg = "Set the {} environment variable".format(setting)
         raise ImproperlyConfigured(error_msg)
-    
+
 def extract_keywords(text):
     # 요청할 URL
     url = 'https://wordcount.com/api/extract_keywords'
@@ -21,7 +21,7 @@ def extract_keywords(text):
 
     with open(secret_file) as f:
         secrets = json.loads(f.read())
-        
+
     # 헤더 설정
     headers = get_secret('headersKeyword', secrets)
 
@@ -47,17 +47,16 @@ def extract_keywords(text):
 
     return keywords
 
-def add_tags_to_db():
-    # SQLite 데이터베이스 경로
-    db_path = 'room_lists.db'
-    conn = sqlite3.connect(db_path)
+def add_tags_to_db(secrets):
+    # MySQL 데이터베이스에 연결
+    conn = connect_db(secrets)
     cursor = conn.cursor()
 
     # 기존 테이블에 'tag' 열 추가
     try:
         cursor.execute("ALTER TABLE dataAnalyze_productTag ADD COLUMN tag TEXT")
-    except sqlite3.OperationalError as e:
-        print(f"Column 'tag' already exists: {e}")
+    except pymysql.MySQLError as e:
+        print(f"Column 'tag' already exists or another error occurred: {e}")
 
     # 모든 행에 대해 태그를 생성하고 업데이트
     cursor.execute("SELECT id, title FROM dataAnalyze_product")
@@ -65,9 +64,9 @@ def add_tags_to_db():
 
     for row in rows:
         row_id, title = row
-        tags = extract_keywords(title)
+        tags = ', '.join(extract_keywords(title))  # 태그를 쉼표로 구분된 문자열로 변환
         print(tags)
-        cursor.execute("UPDATE dataAnalyze_productTag SET tag = ? WHERE id = ?", (tags, row_id))
+        cursor.execute("UPDATE dataAnalyze_productTag SET tag = %s WHERE id = %s", (tags, row_id))
 
     conn.commit()
     conn.close()
@@ -75,4 +74,9 @@ def add_tags_to_db():
     print("Tags have been added to the 'tag' column in the database.")
 
 if __name__ == "__main__":
-    add_tags_to_db()
+    base_dir = os.path.dirname(os.path.abspath(__file__))
+    secret_file = os.path.join(base_dir, '..', '..','secret.json')
+
+    with open(secret_file) as f:
+        secrets = json.loads(f.read())
+    add_tags_to_db(secrets)
