+++ RoomCatcher_Data_AI/dataAnalyze/ver1/transpose_location_to_address_dabang.py
@@ -2,7 +2,7 @@ import requests
 import os 
 import json
 from django.core.exceptions import ImproperlyConfigured
-from .product_crawling import product_crawling
+from dataAnalyze.ver1.product_crawling import product_crawling
 import sqlite3
 
 # 주소 변환 함수
